Fix non-concept builds
