+++ test/unit/include/test_unit/test_any_connection.hpp
@@ -5,8 +5,8 @@
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 //
 
-#ifndef BOOST_MYSQL_TEST_UNIT_INCLUDE_TEST_UNIT_TEST_CONNECTION_HPP
-#define BOOST_MYSQL_TEST_UNIT_INCLUDE_TEST_UNIT_TEST_CONNECTION_HPP
+#ifndef BOOST_MYSQL_TEST_UNIT_INCLUDE_TEST_UNIT_TEST_ANY_CONNECTION_HPP
+#define BOOST_MYSQL_TEST_UNIT_INCLUDE_TEST_UNIT_TEST_ANY_CONNECTION_HPP
 
 #include <boost/mysql/any_connection.hpp>
